requirements.txt
Flask >= 0.10.1
Flask-OAuthlib
Flask-Login
Flask-JWT
coveralls
peewee >= 2.4.2
lupa >= 1.0.1
python-geoip-geolite2
marshmallow
marshmallow_jsonapi
geoip2
statsd
uritemplate.py
requests
pymysql
pyflakes
pytest-mock
pytest-cov
git+https://github.com/FAForever/faftools.git@develop#egg=faftools
PyJWT
